% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chi_o.r
\name{calculate_chi_o}
\alias{calculate_chi_o}
\title{Calculation of chi_o}
\usage{
calculate_chi_o(data, ColPhotos, ColVPD, ColTair, C, Z)
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables.}

\item{ColPhotos}{column name of numeric vector containing time series of photosynthesis data (umol CO2 m-2 s-1)}

\item{ColVPD}{column name of numeric vector containing time series of vapor pressure deficit (kPa).}

\item{ColTair}{column name of numeric vector containing time series of air temperature (deg C).}

\item{C}{Empirical coeficient for C3 species.}

\item{Z}{Z- numeric value defining elevation (km).}
}
\value{
a numeric value:
        \item{chi_o}{long-term effective "internal" leaf-to-ambient CO2 (r.u)}
}
\description{
Calculation of chi_o (see Wang et al., 2017; Plant Nature).
}
\details{
the following metrics are calculated:

         logistic_chi_o:

           \deqn{logistic_chi_o = 0.0545*(Tair_g-25)-0.58*log(VPD_g)-0.0815*Z+C}

           \deqn{chi_o <- exp(logistic_chi_o)/(1+exp(logistic_chi_o))}

         Tair_g and VPD_g are calculated based on the mean value of the growing period.
         The growing period is estimated as those periods over the 85 quantile of Photos.
}
\note{
chi_o is unitless.
      Photos is calculated based on night-time NEE partitioning approach (Reichstein et al., 2005).
}
\examples{
## filter data for dry periods and daytime at DE-Tha in June 2014

calculate_chi_o(data= EddySample
,ColPhotos = "GPP_NT_VUT_MEAN"
,ColVPD = "VPD_F"
,ColTair = "TA_F"
,C = 1.189 ##<< Empirical coeficient for C3 species (see Wang et al., 2017; Plant Nature).
,Z=0.27)
}
\references{
Wang, H., I. C. Prentice, et al., (2017), Towards a universal model
            for carbon dioxide uptake by plants, Nature Plants, 3(9), 734-741.

            Reichstein, M., et al. (2005), On the separation of net ecosystem exchange
            into assimilation and ecosystem respiration: review and improved algorithm,
            Global Change Biology, 11(9), 1424-1439.
}
