% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.r
\name{cost_function}
\alias{cost_function}
\title{Multi-objective optimality-based cost function}
\usage{
cost_function(par, data, ColPhotos, ColPhotos_unc, ColH, ColVPD, ColTair,
  ColPair, ColQ, ColCa, ColUstar, ColWS, ColSW_in, Chi_o, WUE_o)
}
\arguments{
\item{par}{A vector containing 4 parameters (a1,Do,To,beta)}

\item{data}{Data.frame or matrix containing all required variables.}

\item{ColPhotos}{Column name of numeric vector containing time series of photosynthesis data (umol CO2 m-2 s-1).}

\item{ColPhotos_unc}{Column name of numeric vector containing time series of photosynthesis uncertainties (umol CO2 m-2 s-1).}

\item{ColH}{Column name of numeric vector containing time series of sensible heat flux (W m-2).}

\item{ColVPD}{Column name of numeric vector containing time series of vapor pressure deficit (hPa).}

\item{ColTair}{Column name of numeric vector containing time series of air temperature (deg C).}

\item{ColPair}{Column name of numeric vector containing time series of atmospheric pressure (kPa).}

\item{ColQ}{Column name of numeric vector containing time series of photosynthetic active radiation (umol m-2 s-1).}

\item{ColCa}{Column name of numeric vector containing time series of atmospheric CO2 concentration (umol Co2 mol air-1).}

\item{ColUstar}{Column name of numeric vector containing time series of wind friction velocity (m s-1).}

\item{ColWS}{Column name of numeric vector containing time series of wind velocity (m s-1).}

\item{ColSW_in}{Column name of numeric vector containing time series of incoming short-wave radiation (W m-2).}

\item{Chi_o}{Long-term effective chi}

\item{WUE_o}{Long-term effective WUE}
}
\value{
a numeric value:
        \item{OF}{the summed cost to minimize both the mismatch between observed and modeled Photos and the unit cost of transpiration}
}
\description{
Cost function to be implemented in an optimization routine of the ETpartitioning package (see Perez-Priego et al., 2018).
}
\details{
the multi-objective function is defined as:


           \deqn{OF <- sum((photos-photosy_mod)/photos_unc)^2)/n + phi}


         where phi invokes optimality theory by minimizing the following term

         \deqn{phi <- (sum(transpiration_mod)/sum(photos_mod)*WUE_o}
}
\note{
The 4 model parameters (a1, Do, Topt and beta, see Perez-Priego et al., 2018) are estimated using a multi-constraint Markov Chain Monte Carlo (MCMC).The objective function (OF) is to find those numerical solutions that minimize not only the mismatch between observed and modeled Photos but also the unit cost of transpiration by introducing a conditional factor demand (phi), which invokes the optimality hypothesis. The phi term is to be defined as the integrated cost of transpiration (i.e. transpiration_mod/photos_mod) over a time period (5 days) normalized by a factor describing the long-term effective water use efficiency (WUE_o).
}
\examples{
 ## Selecting a single day (e.g. 15-05-2011)
 tmp <-  EddySample[ EddySample$TIMESTAMP_START>  201105150000,]
 tmp <-  tmp[tmp$TIMESTAMP_START<  201105160000,]
 ## Defining parameter values
 par <- c(200, 0.2, 25, 0.6)

cost_function(
 data=tmp
 ,par=par
 ,ColPhotos="GPP_NT_VUT_MEAN"
 ,ColPhotos_unc ="NEE_VUT_USTAR50_JOINTUNC"
 ,ColH="H_F_MDS"
 ,ColVPD="VPD_F"
 ,ColTair="TA_F"
 ,ColPair="PA_F"
 ,ColQ="PPFD_IN"
 ,ColCa="CO2_F_MDS"
 ,ColUstar="USTAR"
 ,ColWS="WS_F"
 ,ColSW_in="SW_IN_F"
 ,Chi_o = 0.88
 ,WUE_o = 24.25
)
}
\references{
Perez-Priego, O., G. Katul, M. Reichstein et al. Partitioning eddy covariance
            water flux components using physiological and micrometeorological approaches,
            Journal of Geophysical Research: Biogeosciences. In press


            Reichstein, M., et al. (2005), On the separation of net ecosystem exchange
            into assimilation and ecosystem respiration: review and improved algorithm,
            Global Change Biology, 11(9), 1424-1439.
}
