% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_model.r
\name{gc_model}
\alias{gc_model}
\title{Calculation of gc}
\usage{
gc_model(par, data, Q, VPD, Tair, gcmax)
}
\arguments{
\item{par}{Set of parameter for the respective sensitivity funciton.}

\item{data}{Data.frame or matrix containing all required variables.}

\item{Q}{Vector containing time series of photosynthetic active radiation (umol CO2 m-2 s-1)}

\item{VPD}{Vector containing time series of vapor pressure deficit (kPa).}

\item{Tair}{Vector containing time series of air temperature (deg C).}

\item{gcmax}{Empirical parameter defining maximum conductance (m s-1 or mol m-2 s-1).}
}
\value{
a numeric value:
        \item{gc_model}{canopy leaf conductance (untis refer to that given by gmax)}
}
\description{
Calculation of canopy stomatal conductance using Jarvis's approach.
}
\details{
the following metrics are calculated:


           \deqn{gc_model <- gcmax*f(Q)*f(VPD)*f(Tair)}
}
\examples{

 ## Selecting a single day (e.g. 15-05-2011)
 tmp <-  EddySample[ EddySample$TIMESTAMP_START>  201105150000,]
 tmp <-  tmp[tmp$TIMESTAMP_START<  201105160000,]
gc_model(par=c(200, 0.2, 25)
,data= tmp
,Q = "PPFD_IN"
,VPD = "VPD_F"
,Tair = "TA_F"
,gcmax = 1)
}
\references{
Perez-Priego, O., G. Katul, M. Reichstein et al. Partitioning eddy covariance
            water flux components using physiological and micrometeorological approaches,
            Journal of Geophysical Research: Biogeosciences. In press
}
