% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpiration_model.r
\name{transpiration_model}
\alias{transpiration_model}
\title{Transpiration model using plant optimization theory.}
\usage{
transpiration_model(par, data, ColPhotos, ColH, ColVPD, ColTair, ColPair,
  ColQ, ColCa, ColUstar, ColWS, ColSW_in, Chi_o, WUE_o)
}
\arguments{
\item{par}{A vector containing 4 parameters (a1,Do,To,beta)}

\item{data}{Data.frame or matrix containing all required variables.}

\item{ColPhotos}{Column name of numeric vector containing time series of photosynthesis data (umol CO2 m-2 s-1).}

\item{ColH}{Column name of numeric vector containing time series of sensible heat flux (W m-2).}

\item{ColVPD}{Column name of numeric vector containing time series of vapor pressure deficit (hPa).}

\item{ColTair}{Column name of numeric vector containing time series of air temperature (deg C).}

\item{ColPair}{Column name of numeric vector containing time series of atmospheric pressure (kPa).}

\item{ColQ}{Column name of numeric vector containing time series of photosynthetic active radiation (umol m-2 s-1).}

\item{ColCa}{Column name of numeric vector containing time series of atmospheric CO2 concentration (umol Co2 mol air-1).}

\item{ColUstar}{Column name of numeric vector containing time series of wind friction velocity (m s-1).}

\item{ColWS}{Column name of numeric vector containing time series of wind velocity (m s-1).}

\item{ColSW_in}{Column name of numeric vector containing time series of incoming short-wave radiation (W m-2).}

\item{Chi_o}{Long-term effective chi}

\item{WUE_o}{Long-term effective WUE}
}
\value{
a numeric vector containing estimates of transpiration rates:
}
\description{
The model provide optimal estimates of transpiration rates using eddy covariance data.
}
\details{
The transpiration is estimated according to gas-difusion equations:


           \deqn{transpiration_mod <-  gw_bulk*VPD_plant/Pair}


         where gw_bulk represents the "bulk" surface conductance and accounts for the influence of stomata and aerodynamic properties.
}
\note{
The parameters (a1, Do, Topt and beta, see Perez-Priego et al., 2018) are estimated using a multi-constraint Markov Chain Monte Carlo (MCMC).The algortihm searches for those optimal solutions that maximize water use efficiency and according to the plant optimization theory.
}
\examples{
 ## Selecting a single day (e.g. 15-05-2011)
 tmp <-  EddySample[ EddySample$TIMESTAMP_START>  201105150000,]
 tmp <-  tmp[tmp$TIMESTAMP_START<  201105160000,]
 ## Defining parameter values
 par <- c(200, 0.2, 25, 0.6)

transpiration_model(
 par=par
 ,data=tmp
 ,ColPhotos="GPP_NT_VUT_MEAN"
 ,ColH="H_F_MDS"
 ,ColVPD="VPD_F"
 ,ColTair="TA_F"
 ,ColPair="PA_F"
 ,ColQ="PPFD_IN"
 ,ColCa="CO2_F_MDS"
 ,ColUstar="USTAR"
 ,ColWS="WS_F"
 ,ColSW_in="SW_IN_F"
 ,Chi_o = 0.88
 ,WUE_o = 24.25
)
}
\references{
Perez-Priego, O., G. Katul, M. Reichstein et al. Partitioning eddy covariance
            water flux components using physiological and micrometeorological approaches,
            Journal of Geophysical Research: Biogeosciences. In press.
}
